+++ pkg/types/types.go
@@ -2,10 +2,7 @@ package types
 
 import (
 	"context"
-	"io"
 )
-
-type ICloser io.Closer
 
 type IRunner interface {
 	Run(context.Context) error